Added test_ft_transfer
