+++ pure-eth-connector/rust-toolchain
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "nightly"
+channel = "nightly-2021-03-25"
 targets = ["wasm32-unknown-unknown"]
